Configure DoD banner on graphical logon screen.
I also included a shell script to do this if I decide in the future it’s cleaner to copy a script, execute it, then remove it.

I also included a handler that “restarts GDM”. The gem-restart command was removed in CentOS 6 so the only way to make the changes appear immediately after these tasks are run is to flop between run levels or press Ctrl+Alt+Backspace from the target host. The bad thing about this is that it kicks out active users with zero warning, so this may be removed in the future.