Add another finding ID to task that copies auditd.conf
Uninstall tftp-server

Add ‘unauthorized_packages’ tag.

It would be more efficient to iterate through a list of unauthorized packages, however this comes at the cost of execution control. If you want to only remove three of the five unauthorized packages because two of them are required in your environment, this would not be possible if all five packages were in a list inside a single task. Open to feedback on a better way to solve this.