Careful testing and conditional removal of tftp-server
Variable rhel6stig_tftp_required variable determines whether these task get executed. The first task checks for the presence of the tftp-server package. Based on the rc of that task, the service will be disabled using a command task. Finally, the package is removed if not needed.

It would be cleaner to just remove the package if not required, thus eliminating the need to disable it. It's a bit silly to disable the service immediiately before removing the package, I admit, but with this first pass at the role I want to address every STIG item individually. Later I'll take a breath, step back, and look at ways to reorganized and consolidate.