Don't ignore errors on tasks that run gconftool-2
These tasks should only be run on a system that has a graphical environment, specified with the variable rhel6stig_xwindows_required. The only time these would fail is if they're run on a system with a graphical environment or if gconftool-2 is not installed. In either case, the play should fail.