Much better rsh, rlogin, and rexec remediation
Check for existance of package. If package exits, check enabled status service. Turn the service off if it is found to be enabled.

While this was fun to work out, the last task in the sequence removes the rsh-server package, rendering the disabling of the services moot. Gotta love the STIGs.