Better handling of TFTP service
Only run tasks if TFTP is in the list of enabled services and not required on the system.

The service module does not properly disable tftp, which is an xinetd servec, so use the command module instead. This is less than ideal. Will fix if I find a better soultion in the future.