Refactor CAT II Tasks (#57)
* Cleanup trailing spaces

* Use proper Jinja filter and adjust comparison

* Correct variable name used for audit task

The registered variable name and the variable used the conditional statement did not match.

* Add task to remediate all of finding V-38501

* Improve formatting of task title and add tags

* Quote shell tasks

* Add extra newline to tasks with multi-line name

* Ignore errors on audit task

Remov task that checks for LDAP. May no longer be needed based on new AUDIT tasks.

* Correct syntax error sysctl task

* Only modify the ypbind and cron services if they are installed

* Better handling of TFTP service

Only run tasks if TFTP is in the list of enabled services and not required on the system.

The service module does not properly disable tftp, which is an xinetd servec, so use the command module instead. This is less than ideal. Will fix if I find a better soultion in the future.

* Add validate to tasks that change sshd_config

This is to prevent sshd from failing to reload if a bad argument is inserted into the config file.

* Add sshd_config tasks

Add a validate statement to the patch tasks to avoid breaking ssh if the synax in the file is incorrect.

Remove tasks using old syntax, particularly the ssh tasks that resulted in an invalid sshd_config.

* Add sysctl_set to sysctl tasks

* Change logic around IPv6 tasks

Change variable rhel6stig_ipv6_in_use to rhel6stig_ipv6_required since it is now
manually set and is not changed using set_fact like before. This allows explicit
control over whether or not IPv6 is enabled or disabled.

Update all tasks and README with new variable name.

Remove tasks from prelim.yml that check for IPv6 addresses.

Delet files/disable-ipv6.conf since it is created by a lineinfile task and no longer copied.

Change the linenfile task that disables IPv6 so that it can also enable it if rhel6stig_ipv6_required is set to True.

* Correct audit variable name

* Do not add ipv6 sysctl settings if IPv6 is disabled and igoner errors.

This fixes #55

* Remove duplicate V-38484 task

* Refactor V-38615

* Refactor V-38617

* Refactor V-38619

* Refactor V-38620 and V-38621

* Refactor V-38622

* Move task that lists rsyslog files to prelim.yml

Since it used by mulitple checks, this qualifies as a preliminary task.

* Refactor V-38623

* Refactor V-38624

* Move V-38625 to not_automated.yml

The corrective action for this changed in R11 because implementing TLS in LDAP is complex. It used to be a one line change to /etc/pam_ldap.conf but now it states "Enable TLS with LDAP".

Helps with issue #48

* Add task to verify OS and version

Role will fail if run against any other OS than RHEL/CentOS 6.

* Always run prelim_tasks.yml

This is more reliable than having to remember to run with the prelim_tasks tag.

* Update meta information

- use galaxy_tags instead of categories
- expand author information
- increase minimum Ansible version to 2.0

* Refactor V-38626

* Add V-38627

* Refactor V-38628

* Refactor V-38629

* Refactor V-38630

* Add V-38633

* Reformat handlers

* Make default variable comments easier to read

* Add V-38634

* Add V-38636

* Correct type-o in V-38468 name

* Refactor V-38637

* Refactor V-38638

* Refactor V-38643

* Refactor V-38652 and V-38654

* Add V-38658

* Refactor V-38660

* Refactor V-38663

* Correct tag for V-38546 patch task

* Correctly name audit variables

* Refactor V-38664 and V-38665

* Refactor V-38670, V-38673, V-38695, V-38696, V-38698, and V-38700

* Refactor V-38671 and V-38647