BUGFIX: (#65)
Fixed issue where the backend command `sysctl -p` was throwing an error
about unknown keys. This is due to the particular kernal modules not
being loaded but is also not a fatal thing. Added `ignoreerrors` to the
V-38597 task as is the case with the other invocations of sysctl module