dangerous regexp whitespace error, only match space/tab (#127)
The \s character set also matches newlines.  If nullok is the last
thing on a line, the following line is effectively removed by being
merged into the current line.  If this next line happens to be the
pam_deny.so line, auth always returns success.  The proper thing to
do here is to only match space and tab characters explicitly.